Finish task without validations

+++ Pages/SearchHistory.cshtml.cs
@@ -2,10 +2,12 @@ using LeapYearApp.Data;
 using LeapYearApp.Extensions;
 using LeapYearApp.Models.Domain;
 using LeapYearApp.Repositories;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
+using System.Security.Claims;
 
 namespace LeapYearApp.Pages
 {
@@ -14,13 +16,18 @@ namespace LeapYearApp.Pages
         private readonly LeapYearAppDbContext _leapYearAppDbContext;
         private readonly IConfiguration _configuration;
         private readonly IYearNameFormRepository _yearNameFormRepository;
+        private readonly SignInManager<IdentityUser> _signInManager;
+        private readonly UserManager<IdentityUser> _userManager;
+
         public PaginatedList<YearNameForm> YearNameForms { get; set; }
 
-        public SearchHistoryModel(LeapYearAppDbContext leapYearAppDbContext, IConfiguration configuration, IYearNameFormRepository yearNameFormRepository)
+        public SearchHistoryModel(LeapYearAppDbContext leapYearAppDbContext, IConfiguration configuration, IYearNameFormRepository yearNameFormRepository, SignInManager<IdentityUser> signInManager, UserManager<IdentityUser> userManager)
         {
             _leapYearAppDbContext = leapYearAppDbContext;
             _configuration = configuration;
             _yearNameFormRepository = yearNameFormRepository;
+            _signInManager = signInManager;
+            _userManager = userManager;
         }
 
         public async Task OnGetAsync(int? pageIndex)
@@ -37,13 +44,29 @@ namespace LeapYearApp.Pages
 
         public async Task<IActionResult> OnPostDelete(Guid id)
         {
-            var deleted = await _yearNameFormRepository.DeleteAsync(id);
+            var yearNameForm = await _yearNameFormRepository.GetByIdAsync(id);
 
-            if (deleted)
+            bool isUserSignedIn = _signInManager.IsSignedIn(User);
+
+            if (!isUserSignedIn)
             {
                 return RedirectToPage("SearchHistory");
             }
-            return Page();
+            else
+            {
+                Guid userId = Guid.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value);
+                IdentityUser user = await _userManager.FindByIdAsync(userId.ToString());
+                if (yearNameForm.UserId == userId && yearNameForm.Login == user.UserName)
+                {
+                    var deleted = await _yearNameFormRepository.DeleteAsync(id);
+
+                    if (deleted)
+                    {
+                        return RedirectToPage("SearchHistory");
+                    }
+                }
+            }
+            return RedirectToPage("SearchHistory");
         }
     }
 }
